requirements.txt
Flask==2.3.4
Flask-Cors==4.0.0
Flask-Sockets==0.2.1
gevent==21.12.0
gevent-websocket==0.10.1
deepgram-sdk==0.6.0
gunicorn==20.1.0